requirements.txt
beautifulsoup4==4.12.2
certifi==2024.8.22
chardet==5.0.0
colorama==0.4.6
cv2==4.8.0
idna==3.4
numpy==1.26.1
opencv-python==4.8.0.76
pyaudio==0.2.12
pyautogui==0.9.54
pyttsx3==2.90
pywikihow==0.5.1
requests==2.31.0
speedtest-cli==2.11.1
SpeechRecognition==3.10.0
wikipedia==1.4.0
pywhatkit==5.4
